Add a `TextField` in the demo and remove an inconsistent `placeholder` argument in `TextField`'s implementation
It turns out that the `placeholder` argument supported only by desktop/Android and the `helperText` supported only by KMDC are not the same thing. The placeholder is shown in the text field when there is no text, while the helper text is shown below the text field when it's in focus.