Initially add the common APIs for `Modifier`, `Box`, `Alignment`, and `BoxScope` without testing them yet
the Compose Multiplatform "ui-unit" module is reused directly on JS without any wrappers created.

The `expect` code follows that in `com.huanshankeji.compose`, in which non-existent APIs in Kobweb Silk are removed. The actual code is implemented in the `androidxCommonMain` source set first and the copied to the `jsMain` source set and adapted there.