Initially implement navigation adapting copied code from the Compose multiplatform navigation library ("org.jetbrains.androidx.navigation:navigation-compose")
The code is called and tested in an internal project. A weired bug occurs that, on Compose HTML / JS DOM, when using destructuring declarations with remembered states and calling them in the `NavHost` `composable`, the state is always reset to the initial value followed by that one update set operation. It works fine when using delegated properties.