Try fixing the bug mentioned in the previous commit without success
It seems that it's a bug in Compose HTML or the Compose compiler on JS rather than this library, or the Compose compiler and runtime in general. I replaced the code in the `androidxCommonMain` source set with the same code in `jsMain` and test it in an internal project, and it turns out that the problem doesn't occur on JVM desktop.

The reason for the bug is that the change in the state change is not propagated and the first destructured component, aka the value, is treated as a static one. After experimenting by printing changed parameters in `LaunchedEffect`, it turns out that every time the state is mutated, on both JVM and JS DOM, the `builder` parameter in the first `NavHost` overload is changed and leads to recomposition but the `graph` parameter in the second `NavHost` overload is not. The change of `builder` doesn't trigger the change of `graph`, which I suspect is due to some skipping mechanism.

I also tried the following approaches, all of which didn't work:
1. removing the `remember` in the `NavHost` with `builder`
1. adding `@NonSkippableComposable` to both `NavHost`s
1. making the `NavHost` with `graph` inline
1. wrapping `builder` in a `@DontMemoize` lambda
1. make the `builder` an interface (and also a SAM functional interface (interestingly, on JS it crashes failing to call the SAM method, which is another issue))

I also tried the following approaches on the state in an internal project:
1. replacing the destructuring declarations with accessing the mutable value property directly and storing the value and the setter in separate `val`s, didn't work
1. using the state's `value` directly in the lambda, worked

So assuming the behavior on JVM desktop is correct, the bug probably lies in that the `val` property destructured from the state used in the lambda is treated as an immutable property in a closure, and a lambda is considered stable by the Compose compiler so not updated. While on JVM desktop the compiler adds some mechanism similar to `rememberUpdatedState` for the lambda, it doesn't do it on JS or Compose HTML. However, it's also weired that the `List` and `LazyColumn` composables adopt a similar approach, but they don't have this issue as I tested.