Add a common `BoxWithConstraints` layout composable
It has been tested manually in an internal project on JVM desktop, but not on JS DOM yet.

Since the APIs provided, especially those members of `BoxWithConstraintsScope`, are a subset of the corresponding `androidx.compose` ones, also consider moving it out of the `ext` package. The counterargument is that `minWidth` and `minHeight` of `androidx.compose.foundation.layout.BoxWithConstraintsScope` seem not possible to be supported on JS DOM, which is the reason why it's still kept in the `ext` package.

Add a `CommonModifier` typealias BTW.