Revert "Bump the Android SDK version to 35"
This reverts commit b2b76bff724f79daa781b2234902904b4210ab0e.

Downgrade the Android SDK version back to 34 as warned by AGP

```text
We recommend using a newer Android Gradle plugin to use compileSdk = 35

This Android Gradle plugin (8.5.2) was tested up to compileSdk = 34.

You are strongly encouraged to update your project to use a newer
Android Gradle plugin that has been tested with compileSdk = 35.

If you are already using the latest version of the Android Gradle plugin,
you may need to wait until a newer version with support for compileSdk = 35 is available.
```